refactor: repository 2개로 분류
