+++ app/src/main/java/woowacourse/shopping/ShoppingRepository.kt
@@ -1,24 +1,10 @@
 package woowacourse.shopping
 
-import woowacourse.shopping.domain.Product
 import woowacourse.shopping.domain.ShoppingCart
 import woowacourse.shopping.domain.ShoppingCartItem
 
 interface ShoppingRepository {
-    fun products(): List<Product>
-
-    fun products(
-        startPosition: Int,
-        offset: Int,
-    ): List<Product>
-
-    fun productById(id: Long): Product
-
-    fun productsTotalSize(): Int
-
-    fun userId(): Long
-
-    fun shoppingCart(userId: Long): ShoppingCart
+    fun shoppingCart(): ShoppingCart
 
     fun shoppingCartItems(
         page: Int,
