+++ app/src/main/java/woowacourse/shopping/util/ViewModelFactory.kt
@@ -2,6 +2,7 @@ package woowacourse.shopping.util
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import woowacourse.shopping.DummyProductRepository
 import woowacourse.shopping.DummyShoppingRepository
 import woowacourse.shopping.productdetail.ProductDetailViewModel
 import woowacourse.shopping.productlist.ProductListViewModel
@@ -11,12 +12,11 @@ class ViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return when {
             modelClass.isAssignableFrom(ProductDetailViewModel::class.java) -> {
-                val repository = DummyShoppingRepository
-                ProductDetailViewModel(repository) as T
+                ProductDetailViewModel(DummyProductRepository, DummyShoppingRepository) as T
             }
 
             modelClass.isAssignableFrom(ProductListViewModel::class.java) -> {
-                val repository = DummyShoppingRepository
+                val repository = DummyProductRepository
                 ProductListViewModel(repository) as T
             }
