+++ app/src/test/java/woowacourse/shopping/viewmodel/ProductDetailViewModelTest.kt
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
+import woowacourse.shopping.DummyProductRepository
 import woowacourse.shopping.DummyShoppingRepository
 import woowacourse.shopping.productdetail.ProductDetailViewModel
 import woowacourse.shopping.viewmodel.fixtures.InstantTaskExecutorExtension
@@ -15,7 +16,7 @@ class ProductDetailViewModelTest {
 
     @BeforeEach
     fun setUp() {
-        viewModel = ProductDetailViewModel(DummyShoppingRepository)
+        viewModel = ProductDetailViewModel(DummyProductRepository, DummyShoppingRepository)
     }
 
     @Test
